requirements.txt
ultralytics
fastapi
uvicorn
python-multipart
opencv-python-headless
pillow
numpy